C2.8.2. Base for Chess app. Base for unit tests part1
