+++ chess/src/main/java/en/exceptions/FigureNotFoundException.java
@@ -1,0 +1,7 @@
+package en.exceptions;
+
+public class FigureNotFoundException extends RuntimeException {
+    public FigureNotFoundException(String msg) {
+            super(msg);
+        }
+    }
