+++ chess/src/main/java/en/exceptions/ImpossibleMoveException.java
@@ -1,0 +1,7 @@
+package en.exceptions;
+
+public class ImpossibleMoveException extends RuntimeException {
+    public ImpossibleMoveException(String msg) {
+            super(msg);
+        }
+    }
