+++ chess/src/main/java/en/exceptions/OccupiedWayException.java
@@ -1,0 +1,7 @@
+package en.exceptions;
+
+public class OccupiedWayException extends RuntimeException {
+    public OccupiedWayException(String msg) {
+            super(msg);
+        }
+    }
