--- chess/src/main/java/en/firuges/Figure.java
@@ -1,17 +1,0 @@
-package en.firuges;
-
-public interface Figure {
-    Cell position();
-
-    Cell[] way(Cell source, Cell dest);
-
-    default String icon() {
-        return String.format(
-                "%s.png", this.getClass().getSimpleName()
-        );
-
-    }
-
-    Figure copy(Cell dest);
-
-}
